--- client/@tomickigrzegorz__react-circular-progress-bar.d.ts
@@ -1,1 +1,0 @@
-declare module "@tomickigrzegorz/react-circular-progress-bar";
